requirements.txt
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0